@@ -1,12 +1,9 @@
 package com.aionemu.geobuilder.cgfData;
 
-import java.util.List;
-
 public class CgfMaterialData {
 
   public String name;
   public int matType, matFlags, materialId;
   public boolean isCollision;
-  public List<Integer> multiMaterialIds;
-
+  public int[] multiMaterialChunkIds;
 }
