@@ -2,7 +2,7 @@ package com.aionemu.geobuilder.entries;
 
 public class BrushEntry {
 
-  public int meshIdx;
+  public int meshIndex;
   public float[] matrix;
   public byte eventType = -1;
   public EntryType type = EntryType.NONE;
