@@ -3,6 +3,6 @@ package com.aionemu.geobuilder.entries;
 public class ObjectEntry {
 
   public float x, y, z, scale, rotX, rotY, rotZ;
-  public int objectId;
+  public int meshIndex;
   public EntryType type = EntryType.NONE;
 }
