@@ -7,5 +7,5 @@ import java.util.List;
 public class BrushLstMeshData {
 
   public List<String> meshFileNames;
-  public List<BrushEntry> brushEntries; // BrushEntry.meshIdx is the index for its name in meshFiles list
+  public List<BrushEntry> brushEntries;
 }
