@@ -7,5 +7,5 @@ import java.util.List;
 public class ObjectMeshData {
 
   public List<String> meshFiles;
-  public List<ObjectEntry> objectEntries; // ObjectEntry.objectId is the index for its name in meshFiles list
+  public List<ObjectEntry> objectEntries;
 }
